% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ipd.R
\name{clean_ipd}
\alias{clean_ipd}
\title{Function to clean old HIS excel-based IPD linelist data (.med) and have in a
usable format.
Not a user-facing function (called within read_msf_data())}
\usage{
clean_ipd(country_file)
}
\arguments{
\item{country_file}{Path to a single IPD file (.med)}
}
\description{
Function to clean old HIS excel-based IPD linelist data (.med) and have in a
usable format.
Not a user-facing function (called within read_msf_data())
}
\author{
Alex Spina
}
