% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_opd.R
\name{clean_opd}
\alias{clean_opd}
\title{Function to clean old HIS excel-based OPD linelist data (.opd) and have in a
usable format.
Not a user-facing function (called within read_msf_data())}
\usage{
clean_opd(country_file)
}
\arguments{
\item{country_file}{Path to a single IPD file (.opd)}
}
\description{
Function to clean old HIS excel-based OPD linelist data (.opd) and have in a
usable format.
Not a user-facing function (called within read_msf_data())
}
\author{
Alex Spina
}
