% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis_ipd_beds.R
\name{dhis_ipd_beds}
\alias{dhis_ipd_beds}
\title{Function to pull IPD weekly bed counts from DHIS output (sourced from pivot)
Not a user-facing function (called within read_msf_data())}
\usage{
dhis_ipd_beds(country_file, wards)
}
\arguments{
\item{country_file}{Path to a single IPD file (.csv exported from DHIS2)}

\item{wards}{Specify which wards want to keep
Default is "Medical" (strict definition without ITFC or maternal)
"Medical extension" includes all medical and ITFC + maternal
"All" also includes surgical}
}
\description{
Function to pull IPD weekly bed counts from DHIS output (sourced from pivot)
Not a user-facing function (called within read_msf_data())
}
\author{
Alex Spina
}
