% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/his_ipd_beds.R
\name{his_ipd_beds}
\alias{his_ipd_beds}
\title{Function to pull IPD weekly bed counts from the MMR tab of old HIS excel-based
tools (.med files)
Not a user-facing function (called within read_msf_data())}
\usage{
his_ipd_beds(country_file)
}
\arguments{
\item{country_file}{Path to a single IPD file (.med)}
}
\description{
Function to pull IPD weekly bed counts from the MMR tab of old HIS excel-based
tools (.med files)
Not a user-facing function (called within read_msf_data())
}
\author{
Alex Spina
}
