% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Function to get the peak weeks from a spectral periodogram}
\usage{
periodogram(x, counts, start_week = c(2014, 1), period = 52, output = "weeks")
}
\arguments{
\item{x}{dataset}

\item{counts}{variable within x containing counts or rates}

\item{start_week}{the first week in your dataset, defaults to c(2014, 1)}

\item{period}{how many units in a year (52 for weeks)}
}
\description{
Function to get the peak weeks from a spectral periodogram
}
\author{
Alex Spina
}
