% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_ipd.R
\name{recode_ipd}
\alias{recode_ipd}
\title{Function to recode old excel-based HIS IPD linelist data to DHIS2 format
Not a user-facing function (called within read_msf_data())}
\usage{
recode_ipd(x, chronic_defs = TRUE)
}
\arguments{
\item{x}{A dataset produced by \link[epitsa:clean_ipd]{epitsa::clean_ipd}\cr function}

\item{chronic_defs}{Logical (TRUE/FALSE) to define if want to stick to the HIS
team definition for chronic conditions (default is TRUE).
This will put all old HIS chronic conditions in to new DHIS2 "Other chronic conditions".
Otherwise (if set to FALSE) leave as "Chronic diseases" to make available for
individually manipulating to combine with DHIS2 data.}
}
\description{
Function to recode old excel-based HIS IPD linelist data to DHIS2 format
Not a user-facing function (called within read_msf_data())
}
\author{
Alex Spina
}
